1.3.3 - audio player is an interface
