+++ main/build.gradle
@@ -6,7 +6,7 @@ plugins {
 }
 
 ext.moduleName = 'lavaplayer'
-version = '1.3.1'
+version = '1.3.3'
 
 dependencies {
   compile 'com.sedmelluq:lavaplayer-common:1.0.6'
