+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/format/AudioPlayerInputStream.java
@@ -1,8 +1,10 @@
 package com.sedmelluq.discord.lavaplayer.format;
 
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
+import com.sedmelluq.discord.lavaplayer.player.DefaultAudioPlayer;
 import com.sedmelluq.discord.lavaplayer.tools.ExceptionTools;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
+import com.sedmelluq.discord.lavaplayer.track.TrackStateListener;
 import com.sedmelluq.discord.lavaplayer.track.playback.AudioFrame;
 
 import javax.sound.sampled.AudioFormat;
@@ -123,10 +125,8 @@ public class AudioPlayerInputStream extends InputStream {
   }
 
   private void notifyTrackStuck() {
-    AudioTrack track = player.getPlayingTrack();
-
-    if (track != null) {
-      player.onTrackStuck(player.getPlayingTrack(), timeout);
+    if (player instanceof TrackStateListener) {
+      ((TrackStateListener) player).onTrackStuck(player.getPlayingTrack(), timeout);
     }
   }
 }
