+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/track/playback/NonAllocatingAudioFrameBuffer.java
@@ -36,7 +36,6 @@ public class NonAllocatingAudioFrameBuffer extends AbstractAudioFrameBuffer {
     int maximumFrameCount = bufferDuration / (int) format.frameDuration() + 1;
     frames = createFrames(maximumFrameCount, format);
     silentFrame = createSilentFrame(format);
-    bridgeFrame = new MutableAudioFrame();
     this.frameBuffer = new byte[format.expectedChunkSize() * maximumFrameCount];
     worstCaseFrameCount = frameBuffer.length / format.maximumChunkSize();
     this.stopping = stopping;
